updated the messaging so it doesn't echo the original question and made ask_agent a function to import into main
