+++ backend/data/Employee_Directory_Internal_Emails.csv
@@ -1,0 +1,61 @@
+Name,Email,Employee ID,Role,Title,Processes Involved
+Nancy Hardin,nancy.hardin@company.com,E1001,HR Coordinator,Compliance Lead,Process 10
+Hector Shaw,hector.shaw@company.com,E1002,Engineer,HR Partner,"Process 2, Process 10, Process 8"
+Julie Cortez,julie.cortez@company.com,E1003,HR Coordinator,Finance Coordinator,"Process 1, Process 6"
+Kevin Griffith,kevin.griffith@company.com,E1004,Contributor,Finance Coordinator,"Process 4, Process 2, Process 9"
+Brittany Wilson,brittany.wilson@company.com,E1005,Contributor,Systems Engineer,Process 2
+Julian Cooper,julian.cooper@company.com,E1006,Project Manager,Business Analyst,Process 8
+Kyle Montoya,kyle.montoya@company.com,E1007,HR Coordinator,Operations Manager,Process 10
+Brandi Hebert,brandi.hebert@company.com,E1008,Project Manager,Business Analyst,Process 5
+Susan Martinez,susan.martinez@company.com,E1009,HR Coordinator,IT Support Specialist,Process 4
+Jeffrey Marquez,jeffrey.marquez@company.com,E1010,HR Coordinator,HR Partner,"Process 9, Process 4"
+Joseph Williams,joseph.williams@company.com,E1011,Team Lead,Software Developer,"Process 10, Process 5, Process 9"
+Thomas Haynes,thomas.haynes@company.com,E1012,Process Owner,IT Support Specialist,"Process 6, Process 2"
+Natasha Walsh,natasha.walsh@company.com,E1013,Contributor,Compliance Lead,"Process 2, Process 6, Process 3"
+Angela Moon,angela.moon@company.com,E1014,Engineer,Business Analyst,"Process 3, Process 2"
+Rita Turner,rita.turner@company.com,E1015,HR Coordinator,Software Developer,"Process 2, Process 6, Process 3"
+Bruce Peters II,bruce.ii@company.com,E1016,Reviewer,Operations Manager,"Process 1, Process 2"
+William Young,william.young@company.com,E1017,Analyst,Systems Engineer,"Process 3, Process 5"
+Zachary Flores,zachary.flores@company.com,E1018,Team Lead,Finance Coordinator,Process 3
+Johnathan Bush,johnathan.bush@company.com,E1019,Engineer,Business Analyst,"Process 3, Process 5, Process 1"
+Anthony Phillips,anthony.phillips@company.com,E1020,Analyst,HR Partner,"Process 6, Process 2"
+Ashley Caldwell MD,ashley.md@company.com,E1021,Project Manager,IT Support Specialist,"Process 4, Process 10, Process 5"
+Timothy Gutierrez,timothy.gutierrez@company.com,E1022,Engineer,Systems Engineer,"Process 9, Process 3, Process 7"
+Steven Gonzalez,steven.gonzalez@company.com,E1023,Reviewer,Compliance Lead,"Process 10, Process 6"
+Ashley Parker,ashley.parker@company.com,E1024,Contributor,Software Developer,Process 3
+Melissa Craig,melissa.craig@company.com,E1025,Contributor,IT Support Specialist,"Process 8, Process 6"
+Mrs. Alexandra Hughes,alexandra.hughes@company.com,E1026,Process Owner,Business Analyst,"Process 7, Process 6"
+Jennifer Barrett,jennifer.barrett@company.com,E1027,Team Lead,HR Partner,"Process 8, Process 7"
+Matthew Wood,matthew.wood@company.com,E1028,Team Lead,Finance Coordinator,"Process 10, Process 8"
+Katrina Duncan,katrina.duncan@company.com,E1029,Process Owner,Business Analyst,Process 10
+Aaron Holmes,aaron.holmes@company.com,E1030,Reviewer,IT Support Specialist,Process 5
+Danny Wilson,danny.wilson@company.com,E1031,Team Lead,Software Developer,Process 5
+Theresa Walker,theresa.walker@company.com,E1032,HR Coordinator,Operations Manager,"Process 3, Process 6, Process 7"
+Samantha Spencer,samantha.spencer@company.com,E1033,Contributor,HR Partner,Process 3
+John Blake,john.blake@company.com,E1034,Process Owner,Systems Engineer,"Process 1, Process 10, Process 6"
+Anita Kim,anita.kim@company.com,E1035,Process Owner,Software Developer,"Process 9, Process 4"
+Michael Collins,michael.collins@company.com,E1036,Team Lead,Business Analyst,"Process 9, Process 4, Process 7"
+Alexandra Robles,alexandra.robles@company.com,E1037,Analyst,Software Developer,"Process 5, Process 9, Process 3"
+Emma Nicholson,emma.nicholson@company.com,E1038,Team Lead,Business Analyst,Process 10
+Darlene Campbell,darlene.campbell@company.com,E1039,Reviewer,HR Partner,Process 9
+Catherine Gregory,catherine.gregory@company.com,E1040,Contributor,Business Analyst,"Process 9, Process 4"
+Gary Gray,gary.gray@company.com,E1041,Contributor,Operations Manager,"Process 8, Process 5, Process 7"
+Carol Hernandez,carol.hernandez@company.com,E1042,Analyst,IT Support Specialist,Process 5
+Antonio Cantrell,antonio.cantrell@company.com,E1043,Process Owner,Finance Coordinator,"Process 2, Process 8"
+Mary Moore,mary.moore@company.com,E1044,Analyst,Software Developer,Process 9
+William Salazar,william.salazar@company.com,E1045,Analyst,HR Partner,Process 9
+Alan Johnson,alan.johnson@company.com,E1046,HR Coordinator,Operations Manager,"Process 9, Process 3, Process 10"
+Eddie Jackson,eddie.jackson@company.com,E1047,HR Coordinator,IT Support Specialist,"Process 4, Process 9"
+Jackie Johnson,jackie.johnson@company.com,E1048,Project Manager,Business Analyst,Process 8
+Philip Salazar,philip.salazar@company.com,E1049,Engineer,Software Developer,"Process 3, Process 8"
+Kristin Adams,kristin.adams@company.com,E1050,Engineer,HR Partner,"Process 1, Process 5"
+Natalie Ortiz,natalie.ortiz@company.com,E1051,Reviewer,IT Support Specialist,Process 9
+Todd Mccoy,todd.mccoy@company.com,E1052,Contributor,Operations Manager,Process 8
+Jesse Gonzalez,jesse.gonzalez@company.com,E1053,Engineer,Systems Engineer,Process 2
+Jason Shah,jason.shah@company.com,E1054,Process Owner,HR Partner,"Process 1, Process 5"
+Jennifer Freeman,jennifer.freeman@company.com,E1055,Analyst,IT Support Specialist,"Process 1, Process 3"
+Robert Rodriguez,robert.rodriguez@company.com,E1056,Process Owner,Systems Engineer,Process 10
+Jeffrey Oconnor,jeffrey.oconnor@company.com,E1057,Contributor,Software Developer,Process 10
+Jeffrey Chase,jeffrey.chase@company.com,E1058,Contributor,Business Analyst,"Process 4, Process 9"
+Joshua Ruiz,joshua.ruiz@company.com,E1059,HR Coordinator,Software Developer,"Process 6, Process 8, Process 7"
+Brian Brown,brian.brown@company.com,E1060,Team Lead,Systems Engineer,"Process 5, Process 3, Process 7"
